# Conflicts:
#	BackgroundLockChanger/project.json
#	LockBe/LockBe.csproj
#	LockBe/Package.appxmanifest
#	LockBe/project.json
#	ShareClass/project.json

+++ ShareClass/Utilities/HttpService.cs
@@ -101,7 +101,7 @@ namespace ShareClass.Utilities
             return null;
         }
 
-        public static async Task<bool> GetHeadTask(string url)
+        public static async Task<bool> GetHeadTask(string url, bool tryGetIfFailed = true, bool callFromRss = false)
         {
             //Check Uri
             Uri uriResult;
@@ -111,6 +111,20 @@ namespace ShareClass.Utilities
                 return false;
             }
 
+            if (!(uriResult.Scheme.ToLower() == "http" || uriResult.Scheme.ToLower() == "https"))
+            {
+                return false;
+            }
+
+            var tempStr = url;
+            var count = 0;
+            while (tempStr.IndexOf("//", StringComparison.Ordinal) != -1)
+            {
+                count++;
+                tempStr = tempStr.Remove(tempStr.IndexOf("//", StringComparison.Ordinal), 2);
+            }
+            if (count > 1) return false;
+
             HttpClient httpClient = new HttpClient();
             HttpRequestMessage request = new HttpRequestMessage
             {
@@ -118,14 +132,32 @@ namespace ShareClass.Utilities
                 RequestUri = uriResult
             };
 
-            HttpResponseMessage response = await httpClient.SendAsync(request);
+            HttpResponseMessage response;
+
+            try
+            {
+                response = await httpClient.SendAsync(request);
+            }
+            catch (Exception)
+            {
+                if (callFromRss)
+                {
+                    SettingManager.SetRssLink("");
+                }
+                throw;
+            }
+          
 
             if (response.StatusCode != HttpStatusCode.OK)
             {
-                var response2 = await GetResponse(url);
-                return response2.StatusCode == HttpStatusCode.OK;
+                if (tryGetIfFailed)
+                {
+                    var response2 = await GetResponse(url);
+                    return response2.StatusCode == HttpStatusCode.OK;
+                }
+                return false;
             }
-            return false;
+            return true;
         }
 
         public static async Task DownloadImage(string url)
@@ -140,20 +172,30 @@ namespace ShareClass.Utilities
 
             if (!StorageHelper.IsFileExisted(backgroundFolder, fileName))
             {
-                StorageFile file = await
-                    StorageFile.CreateStreamedFileFromUriAsync(fileName, uri,
-                        RandomAccessStreamReference.CreateFromUri(uri));
-
-                var fileList = await backgroundFolder.GetFilesAsync();
-                foreach (StorageFile storageFile in fileList)
+                //Check if link is valid
+                bool isLinkValid = await GetHeadTask(url, false);
+                if (isLinkValid)
                 {
-                    if (StorageHelper.IsFileExisted(backgroundFolder, storageFile.Name))
-                    {
-                        await storageFile.DeleteAsync(StorageDeleteOption.Default);
-                    }
-                }
+                    StorageFile file = await
+                        StorageFile.CreateStreamedFileFromUriAsync(fileName, uri,
+                            RandomAccessStreamReference.CreateFromUri(uri));
 
-                await file.CopyAsync(backgroundFolder);
+                    var fileList = await backgroundFolder.GetFilesAsync();
+                    foreach (StorageFile storageFile in fileList)
+                    {
+                        if (StorageHelper.IsFileExisted(backgroundFolder, storageFile.Name))
+                        {
+                            await storageFile.DeleteAsync(StorageDeleteOption.Default);
+                        }
+                    }
+
+                    await file.CopyAsync(backgroundFolder);
+                }
+                else
+                {
+                    //Get the current file from backgroundFolder
+                    //by doing nothing here
+                }
             }
             else
             {
@@ -177,7 +219,6 @@ namespace ShareClass.Utilities
                 }
 
                 await file.CopyAsync(backgroundFolder);
-
             }
         }
     }
